+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-#1.1
-###Names
-* Delete implementation Decimal from BigDecimal
+##1.1
+* Delete implementation Decimal from BigDecimal;
+
+##1.2
+* Added a converter for the Decimal class for use with the JPA Repository;
+* Added .toString() method fot the Decimal class;